// src/services/api/generated/types.gen.ts
export type LocalDateTime = string;

export type ProjectCreateRequest = {
    name: string;
};

export type ProjectInfoResponse = {
    id?: Uuid;
    name?: string;
    createdAt?: LocalDateTime;
    updatedAt?: LocalDateTime;
    documentCount?: number;
};

export type ProjectUpdateRequest = {
    name: string;
};

export type SearchRequest = {
    query: string;
    projectId: Uuid;
};

export type SearchResponse = {
    results?: Array<SearchResult>;
};

export type SearchResult = {
    id?: Uuid;
    chunkText?: string;
    chunkIndex?: number;
    source?: string;
    distance?: number;
};

export type TextRequest = {
    text: string;
    projectId: Uuid;
};

export type Uuid = string;